+++ N1/drivers/adc.h
@@ -6,6 +6,6 @@
 #include <avr/io.h>
 
 int ADC_Init(void);
-uint8_t ADC_Read(char ch);
+int ADC_Read(char ch);
 
 #endif /* ADC_H_ */
